Remove rest of deploy-phase common
