+++ handel/src/common/ecs-containers.ts
@@ -19,7 +19,6 @@ import { deployPhase } from 'handel-extension-support';
 import * as _ from 'lodash';
 import { FargateServiceConfig } from '../services/ecs-fargate/config-types';
 import { EcsServiceConfig } from '../services/ecs/config-types';
-import * as deployPhaseCommon from './deploy-phase-common';
 import * as routingSection from './ecs-routing';
 import { ContainerConfig, HandlebarsEcsTemplateContainer } from './ecs-shared-config-types';
 import * as volumesSection from './ecs-volumes';
