+++ handel/src/services/codedeploy/iam-roles.ts
@@ -16,7 +16,6 @@
  */
 import { DeployContext, ServiceContext } from 'handel-extension-api';
 import { deployPhase, handlebars } from 'handel-extension-support';
-import * as deployPhaseCommon from '../../common/deploy-phase-common';
 import * as util from '../../common/util';
 import { CodeDeployServiceConfig } from './config-types';
 
@@ -28,7 +27,6 @@ export async function getStatementsForInstanceRole(ownServiceContext: ServiceCon
         handelBucketName: deployPhase.getHandelUploadsBucketName(accountConfig)
     };
     const compiledPolicyStatements = await handlebars.compileTemplate(ownPolicyStatementsTemplate, handlebarsParams);
-    let ownPolicyStatements = JSON.parse(compiledPolicyStatements);
-    ownPolicyStatements = ownPolicyStatements.concat(deployPhaseCommon.getAppSecretsAccessPolicyStatements(ownServiceContext));
-    return deployPhaseCommon.getAllPolicyStatementsForServiceRole(ownPolicyStatements, dependenciesDeployContexts);
+    const ownPolicyStatements = JSON.parse(compiledPolicyStatements);
+    return deployPhase.getAllPolicyStatementsForServiceRole(ownServiceContext, ownPolicyStatements, dependenciesDeployContexts, true);
 }
