+++ handel/src/services/ecs/ecs-service-template.yml
@@ -20,26 +20,22 @@ Resources:
               - "ecs.amazonaws.com"
             Action:
             - "sts:AssumeRole"
-  ServiceRolePolicy:
-    Type: AWS::IAM::Policy
-    Properties:
-      PolicyName: {{serviceRoleName}}
-      Roles:
-      - !Ref ServiceRole
-      PolicyDocument:
-        Version: "2012-10-17"
-        Statement:
-        - Effect: 'Allow'
-          Action:
-          - "ec2:AuthorizeSecurityGroupIngress"
-          - "ec2:Describe*"
-          - "elasticloadbalancing:DeregisterInstancesFromLoadBalancer"
-          - "elasticloadbalancing:DeregisterTargets"
-          - "elasticloadbalancing:Describe*"
-          - "elasticloadbalancing:RegisterInstancesWithLoadBalancer"
-          - "elasticloadbalancing:RegisterTargets"
-          Resource:
-          - "*"
+      Policies:
+      - PolicyName: {{serviceRoleName}}
+        PolicyDocument:
+          Version: "2012-10-17"
+          Statement:
+          - Effect: 'Allow'
+            Action:
+            - "ec2:AuthorizeSecurityGroupIngress"
+            - "ec2:Describe*"
+            - "elasticloadbalancing:DeregisterInstancesFromLoadBalancer"
+            - "elasticloadbalancing:DeregisterTargets"
+            - "elasticloadbalancing:Describe*"
+            - "elasticloadbalancing:RegisterInstancesWithLoadBalancer"
+            - "elasticloadbalancing:RegisterTargets"
+            Resource:
+            - "*"
 
   #
   # Configure IAM resources for ECS resources
