+++ handel/src/services/kms/index.ts
@@ -17,7 +17,6 @@
 import { DeployContext, PreDeployContext, ServiceConfig, ServiceContext, UnDeployContext } from 'handel-extension-api';
 import { awsCalls, deletePhases, deployPhase, handlebars, tagging } from 'handel-extension-support';
 import * as winston from 'winston';
-import * as deployPhaseCommon from '../../common/deploy-phase-common';
 import {KmsServiceConfig} from './config-types';
 
 const SERVICE_NAME = 'KMS';
