+++ handel/src/services/s3/index.ts
@@ -18,7 +18,6 @@ import { DeployContext, PreDeployContext, ProduceEventsContext, ServiceConfig, S
 import { awsCalls, deletePhases, deployPhase, handlebars, tagging } from 'handel-extension-support';
 import * as winston from 'winston';
 import * as s3Calls from '../../aws/s3-calls';
-import * as deployPhaseCommon from '../../common/deploy-phase-common';
 import * as s3DeployersCommon from '../../common/s3-deployers-common';
 import { STDLIB_PREFIX } from '../stdlib';
 import { HandlebarsS3Template, S3ServiceConfig, S3ServiceEventConsumer, S3ServiceEventFilterList } from './config-types';
