+++ handel/src/services/sqs/index.ts
@@ -25,7 +25,6 @@ import {
 import { awsCalls, deletePhases, deployPhase, handlebars, tagging } from 'handel-extension-support';
 import * as winston from 'winston';
 import * as sqsCalls from '../../aws/sqs-calls';
-import * as deployPhaseCommon from '../../common/deploy-phase-common';
 import { STDLIB_PREFIX } from '../stdlib';
 import {HandlebarsSqsTemplate, SqsServiceConfig} from './config-types';
