+++ handel/test/services/apigateway/proxy/proxy-passthrough-deploy-type-test.ts
@@ -21,7 +21,6 @@ import 'mocha';
 import * as sinon from 'sinon';
 import config from '../../../../src/account-config/account-config';
 import * as route53 from '../../../../src/aws/route53-calls';
-import * as deployPhaseCommon from '../../../../src/common/deploy-phase-common';
 import {APIGatewayConfig} from '../../../../src/services/apigateway/config-types';
 import * as proxyPassthroughDeployType from '../../../../src/services/apigateway/proxy/proxy-passthrough-deploy-type';
 import { STDLIB_PREFIX } from '../../../../src/services/stdlib';
