+++ handel/test/services/apigateway/swagger/swagger-deploy-type-test.ts
@@ -26,7 +26,6 @@ import { deployPhase } from 'handel-extension-support';
 import 'mocha';
 import * as sinon from 'sinon';
 import config from '../../../../src/account-config/account-config';
-import * as deployPhaseCommon from '../../../../src/common/deploy-phase-common';
 import { APIGatewayConfig } from '../../../../src/services/apigateway/config-types';
 import * as swaggerDeployType from '../../../../src/services/apigateway/swagger/swagger-deploy-type';
 import { STDLIB_PREFIX } from '../../../../src/services/stdlib';
