+++ handel/test/services/cloudwatchevent/cloudwatchevent-test.ts
@@ -29,7 +29,6 @@ import 'mocha';
 import * as sinon from 'sinon';
 import config from '../../../src/account-config/account-config';
 import * as cloudWatchEventsCalls from '../../../src/aws/cloudwatch-events-calls';
-import * as deployPhaseCommon from '../../../src/common/deploy-phase-common';
 import * as cloudWatchEvent from '../../../src/services/cloudwatchevent';
 import { CloudWatchEventsConfig, CloudWatchEventsServiceEventConsumer } from '../../../src/services/cloudwatchevent/config-types';
 import { STDLIB_PREFIX } from '../../../src/services/stdlib';
