+++ handel/test/services/iot/iot-test.ts
@@ -28,7 +28,6 @@ import { awsCalls, deployPhase } from 'handel-extension-support';
 import 'mocha';
 import * as sinon from 'sinon';
 import config from '../../../src/account-config/account-config';
-import * as deployPhaseCommon from '../../../src/common/deploy-phase-common';
 import {
 
 } from '../../../src/datatypes';
