+++ handel/test/services/kms/kms-test.ts
@@ -27,7 +27,6 @@ import { deletePhases, deployPhase } from 'handel-extension-support';
 import 'mocha';
 import * as sinon from 'sinon';
 import config from '../../../src/account-config/account-config';
-import * as deployPhaseCommon from '../../../src/common/deploy-phase-common';
 import * as kms from '../../../src/services/kms';
 import { KmsServiceConfig } from '../../../src/services/kms/config-types';
 import { STDLIB_PREFIX } from '../../../src/services/stdlib';
