+++ handel/test/services/route53zone/route53zone-test.ts
@@ -27,7 +27,6 @@ import { deletePhases, deployPhase } from 'handel-extension-support';
 import 'mocha';
 import * as sinon from 'sinon';
 import config from '../../../src/account-config/account-config';
-import * as deployPhaseCommon from '../../../src/common/deploy-phase-common';
 import * as route53 from '../../../src/services/route53zone';
 import { Route53ZoneServiceConfig } from '../../../src/services/route53zone/config-types';
 import { STDLIB_PREFIX } from '../../../src/services/stdlib';
