+++ handel/test/services/stepfunctions/stepfunctions-test.ts
@@ -28,7 +28,6 @@ import { deletePhases, deployPhase } from 'handel-extension-support';
 import 'mocha';
 import * as sinon from 'sinon';
 import config from '../../../src/account-config/account-config';
-import * as deployPhaseCommon from '../../../src/common/deploy-phase-common';
 import * as util from '../../../src/common/util';
 import { STDLIB_PREFIX } from '../../../src/services/stdlib';
 import * as stepfunctions from '../../../src/services/stepfunctions';
